# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nova Launch API",
    "version": "1.0.0",
    "description": "REST API for the Nova Launch Stellar token deployment platform.",
    "contact": {
      "name": "Nova Launch",
      "url": "https://github.com/Emmyt24/Nova-launch"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://api.nova-launch.app",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/tokens/search": {
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "Search and filter tokens",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or symbol"
          },
          {
            "name": "creator",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "minSupply",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxSupply",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasBurns",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created",
                "burned",
                "supply",
                "name"
              ],
              "default": "created"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TokenRecord"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "cached": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/leaderboard/most-burned": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Tokens with highest burn volume",
        "parameters": [
          {
            "$ref": "#/components/parameters/PeriodParam"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard results"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/leaderboard/most-active": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Tokens with most burn transactions",
        "parameters": [
          {
            "$ref": "#/components/parameters/PeriodParam"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard results"
          }
        }
      }
    },
    "/api/leaderboard/newest": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Recently created tokens",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard results"
          }
        }
      }
    },
    "/api/leaderboard/largest-supply": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Tokens with highest total supply",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard results"
          }
        }
      }
    },
    "/api/leaderboard/most-burners": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Tokens with most unique burners",
        "parameters": [
          {
            "$ref": "#/components/parameters/PeriodParam"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard results"
          }
        }
      }
    },
...
    "/api/campaigns/stats/{tokenId}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Campaign statistics for a token (or global)",
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign stats"
          }
        }
      }
    },
...
    "/api/streams/stats/{address}": {
      "get": {
        "tags": [
          "Streams"
        ],
        "summary": "Stream statistics for an address or global",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream stats"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /auth/verify or POST /auth/refresh"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrev": {
            "type": "boolean"
          }
        }
      },
      "TokenRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "decimals": {
            "type": "integer"
          },
          "totalSupply": {
            "type": "string",
            "description": "BigInt as string"
          },
          "initialSupply": {
            "type": "string",
            "description": "BigInt as string"
          },
          "totalBurned": {
            "type": "string",
            "description": "BigInt as string"
          },
          "burnCount": {
            "type": "integer"
          },
          "metadataUri": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT issued for users with role 'admin' or 'super_admin'."
      }
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "PeriodParam": {
        "name": "period",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "24h",
            "7d",
            "30d",
            "all"
          ],
          "default": "7d"
        }
      },
...
    }
  }
}